requirements.txt
Django==2.1
django-filter==2.0.0
djangorestframework==3.7.7
pytz==2018.9
